+++ optd/repository/api/src/schema/get_all_schemas.rs
@@ -1,25 +1,22 @@
 use sea_orm::{
-    ColumnTrait, Condition, ConnectionTrait, DbErr, EntityTrait, QueryFilter, QuerySelect,
+    ColumnTrait, Condition, ConnectionTrait, DbErr, EntityTrait, QueryFilter, QueryOrder,
+    QuerySelect,
 };
 
 use crate::{
-    api::{
-        schema::{GetSchemaInfo, SchemaInfo},
-        snapshot::SnapshotInfo,
-    },
+    schema::SchemaInfo,
+    snapshot::SnapshotInfo,
     entity::{prelude::Schema, schema},
 };
 
-pub async fn get_schema<C>(
-    info: GetSchemaInfo,
+pub async fn get_all_schema_infos<C>(
     db: &C,
     current_snapshot: &mut SnapshotInfo,
-) -> Result<Option<SchemaInfo>, DbErr>
+) -> Result<Vec<SchemaInfo>, DbErr>
 where
     C: ConnectionTrait,
 {
     let snapshot_id = current_snapshot.snapshot_id;
-
     Schema::find()
         .filter(schema::Column::BeginSnapshot.lte(snapshot_id))
         .filter(
@@ -27,12 +24,12 @@ where
                 .add(schema::Column::EndSnapshot.is_null())
                 .add(schema::Column::EndSnapshot.gt(snapshot_id)),
         )
-        .filter(schema::Column::SchemaId.eq(info.schema_id))
         .select_only()
         .column(schema::Column::SchemaId)
         .column(schema::Column::SchemaUuid)
         .column(schema::Column::SchemaName)
+        .order_by_asc(schema::Column::SchemaName)
         .into_partial_model::<SchemaInfo>()
-        .one(db)
+        .all(db)
         .await
 }
