+++ optd/repository/api/src/table/get_all_tables.rs
@@ -1,58 +1,77 @@
+use std::collections::HashMap;
+
 use sea_orm::{
     ColumnTrait, Condition, ConnectionTrait, DbErr, EntityTrait, QueryFilter, QueryOrder,
 };
 
 use crate::{
-    api::{
-        snapshot::SnapshotInfo,
-        table::{ColumnInfo, GetTableInfo, TableInfo},
-    },
+    snapshot::SnapshotInfo,
+    table::{ColumnInfo, TableInfo},
     entity::{column, prelude::*, table},
 };
 
-pub async fn get_table<C>(
-    info: GetTableInfo,
+pub async fn get_all_table_infos<C>(
     db: &C,
     current_snapshot: &mut SnapshotInfo,
-) -> Result<Option<TableInfo>, DbErr>
+) -> Result<Vec<TableInfo>, DbErr>
 where
     C: ConnectionTrait,
 {
     let snapshot_id = current_snapshot.snapshot_id;
-
-    let Some(table) = Table::find()
+    let tables = Table::find()
         .filter(table::Column::BeginSnapshot.lte(snapshot_id))
         .filter(
             Condition::any()
                 .add(table::Column::EndSnapshot.is_null())
                 .add(table::Column::EndSnapshot.gt(snapshot_id)),
         )
-        .filter(table::Column::TableId.eq(info.table_id))
-        .one(db)
-        .await?
-    else {
-        return Ok(None);
-    };
+        .order_by_asc(table::Column::TableId)
+        .all(db)
+        .await?;
 
+    if tables.is_empty() {
+        return Ok(Vec::new());
+    }
+
+    let table_ids = tables
+        .iter()
+        .map(|table| table.table_id)
+        .collect::<Vec<_>>();
     let columns = Column::find()
-        .filter(column::Column::TableId.eq(table.table_id))
+        .filter(column::Column::TableId.is_in(table_ids))
         .filter(column::Column::BeginSnapshot.lte(snapshot_id))
         .filter(
             Condition::any()
                 .add(column::Column::EndSnapshot.is_null())
                 .add(column::Column::EndSnapshot.gt(snapshot_id)),
         )
+        .order_by_asc(column::Column::TableId)
         .order_by_asc(column::Column::ColumnOrder)
         .all(db)
         .await?;
 
-    Ok(Some(TableInfo {
-        table_id: table.table_id,
-        schema_id: table.schema_id,
-        table_uuid: table.table_uuid,
-        table_name: table.table_name,
-        columns: build_columns(columns)?,
-    }))
+    let mut columns_by_table = HashMap::<i64, Vec<column::Model>>::new();
+    for column in columns {
+        columns_by_table
+            .entry(column.table_id)
+            .or_default()
+            .push(column);
+    }
+
+    Ok(tables
+        .into_iter()
+        .map(|table| {
+            Ok(TableInfo {
+                table_id: table.table_id,
+                schema_id: table.schema_id,
+                table_uuid: table.table_uuid,
+                table_name: table.table_name,
+                columns: build_columns(
+                    columns_by_table.remove(&table.table_id).unwrap_or_default(),
+                )?,
+            })
+        })
+        .collect::<Result<Vec<_>, DbErr>>()?)
 }
 
 fn build_columns(columns: Vec<column::Model>) -> Result<Vec<ColumnInfo>, DbErr> {
