+++ optd/repository/entity/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "optd-repository-entity"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+
+[dependencies]
+sea-orm = { workspace = true }
+arrow-schema = { workspace = true }
+derive_more = { version = "2.1.1", features = ["display"] }
