+++ optd/repository/entity/src/snapshot.rs
