+++ optd/repository/entity/src/table_stats.rs
